finished getting device auth token

+++ t_cloud_client.py
@@ -5,16 +5,20 @@ import asyncio
 import json
 import uuid
 import requests
+import json
+import hashlib
+import pathlib
 
 
-class StreamerClient:
-    def __init__(self):
-        self._token=None
-        self._entry_point = streamer_entry_point
+class TCloudClient:
+    def __init__(self, entrypoint, user=None):
+        self._token = None
+        self._user = user
+        self._entrypoint = entrypoint
         self._headers = {'accept':'application/json',
                         'Authorization': None,
                         'Content-Type': 'application/json'}
-        self.http_client = AsyncHTTPClient()
+        self._http_client = AsyncHTTPClient()
 
     def _get_msg(self, response):
         url = response.request.url
@@ -29,14 +33,14 @@ class StreamerClient:
         """Get the access token from Alpha."""
 
         try:
-            api_endpoint = auth_entry_point
+            url = f'{self._entrypoint}auth/token'
 
             headers = {'X-TCLOUD-SERVICE': 'fm',
                         'Content-Type': 'application/json'}
 
-            body = json.dumps(auth_user)
+            body = json.dumps(self._user)
 
-            response = await self.http_client.fetch(api_endpoint,
+            response = await self._http_client.fetch(url,
                                             raise_error=False,
                                             method='POST',
                                             body=body,
@@ -58,21 +62,63 @@ class StreamerClient:
                 return await fetch()
         return response
     
+    async def _get_device_token(self, _device_id):
+        """Get the access token from Alpha."""
+
+        try:
+            url = f'{self._entrypoint}auth/device/token'
+
+            # data = 'e42c56dbfdb7dbc34ded-f25d-488d-bf68-e822b4f6dd2a0P-69-00-DB-FD-B71586507721'.encode('utf-8')
+            # s = hashlib.sha1()
+            # s.update(data)
+            # h = s.hexdigest()
+            # print(h)
+            
+            body = json.dumps({
+                "auth_digest": "b249445fa3cdfccaf48ed270075b805d2d9c7ecc",
+                "serial_number": "0P-69-00-DB-FD-B7",
+                "timestamp": 1586507721,
+                "uuid": "dbc34ded-f25d-488d-bf68-e822b4f6dd2a"
+                })
+            
+            headers = {k: v for k, v in self._headers.items() if v is not None}
+
+            response = await self._http_client.fetch(url,
+                                            raise_error=True,
+                                            method='POST',
+                                            body=body,
+                                            headers=headers)
+            if response.error is None:
+                self._token = f"Bearer {json.loads(response.body)['msg']['v3']['access_token']}"
+                self._headers['Authorization'] = self._token
+                print(f'Getting token successful.')
+            else:
+                print(f'Getting token failed.')
+
+        except Exception as e:
+            print(f'An error occurred while getting token {e}')
+
+    async def _device_auth_promise_fetch(self, device_id, fetch):
+        response = await fetch()
+        if response.code in (401, 599):
+                await self._get_device_token(device_id)
+                return await fetch()
+        return response
+    
     async def _get_cams(self, device_id):
         """List cameras by device_id."""
 
-        api_endpoint = f'{self._entry_point}{device_id}/cams?bypass_status_check=false'
+        url = f'{self._entrypoint}streamer/{device_id}/cams?bypass_status_check=false'
        
         async def fetch():
-            return await self.http_client.fetch(api_endpoint,
+            return await self._http_client.fetch(url,
                                             raise_error=False,
                                             method='GET',
                                             headers=self._headers)
         return await self._auth_promise_fetch(fetch)
 
-    async def get_cam_id(self, device_id, cam_name):
+    async def _get_cam_id(self, device_id, cam_name):
         """Get the camera by the given device and camera name"""
-        
         
         response = await self._get_cams(device_id)
         cams = self._get_msg(response)
@@ -85,20 +131,44 @@ class StreamerClient:
             print('Can not find any camera by the given camera name')
             return
         return match_cams[0]['cam_uid']
-           
-    def download_video(self, cam_id, device_id, start, duration, filename):
-        """Download video to current folder.
+
+    async def _patch_setting(self,device_id, filename):
+        """Upliad the config of the wheel calibration"""
         
-        :param start: timestamp string eg.'1586495803'.
-        :param duration: unit is second and type is int.
-        :param filename: the name gonna be saved.
+        url = f'{self._entrypoint}event_resource/setting'
+       
+        
+        with open(filename,'r') as f:
+            data = json.load(f)
+        body = json.dumps(data)
+        method = 'PATCH'
+        async def fetch():
+            return await self._http_client.fetch(url,
+                                        raise_error=False,
+                                        method=method,
+                                        body=body,
+                                        headers=self._headers
+                                        )
+        return await self._device_auth_promise_fetch(device_id, fetch)
+             
+    async def download_video(self, device_id, cam_name, start, duration):
+        """Download video
+        
+        :param start: Timestamp string.
+        :param duration: int number, unit seconds
         """
         
-        
         try:
-            print('Video downloading...')
+            filename = f'{cam_name}-{start}-{duration}.mkv'
+            if Path(filename).is_file():
+                print(f'{filename} already exist.')
+                return filename
             
-            url = f'{self._entry_point}{device_id}/cams/{cam_id}/clip?pos={start}&duration={duration}&profile=-1'
+            print('Getting camera id...')
+            cam_id = await self._get_cam_id(device_id, cam_name)
+            
+            print('Video downloading...')
+            url = f'{self._entrypoint}streamer/{device_id}/cams/{cam_id}/clip?pos={start}&duration={duration}&profile=-1'
             with requests.get(url, stream=True, headers=self._headers) as r:
                 r.raise_for_status()
                 with open(filename, 'wb') as f:
@@ -108,9 +178,27 @@ class StreamerClient:
                             # f.flush()
             
             print('done')
+            return filename
         except Exception as e:
             print(e)
-
-
+            return None
+ 
+    async def upload_config(self, device_id, filename):
+        """Upliad the config of the wheel calibration.
+        
+        :param start: Timestamp string.
+        :param duration: int number, unit seconds
+        """
+        
+        if not Path(filename).is_file():
+            print(f"The config {filename} doesn't exist.")
+            return False
+        
+        response = await self._patch_setting(device_id, filename)
+        if not response.error:
+            return True
+        else:
+            print(f'{response.error}')
+            return False
